requirements-complete.txt
# ============================================================================
# D.P. Economy - Complete Requirements (Based on setup.cfg)
# Auto-generated from setup.cfg dependencies
# ============================================================================

# Django Core
Django==5.2.7
django-cors-headers>=4.3.0
django-environ>=0.11.2
polib==1.2.0  # Translation file handling

# Database Adapters
psycopg2-binary>=2.9.9  # PostgreSQL (production)
mysqlclient==2.2.7  # MySQL (alternative)

# REST API
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
drf-spectacular>=0.27.0  # OpenAPI/Swagger docs

# Async Tasks & Background Jobs
celery>=5.3.4
redis>=5.0.1
django-q2>=1.6.0  # Django 5.x compatible
tendo==0.3.0  # Singleton instance control

# File Processing & Export
openpyxl==3.1.5  # Excel read
XlsxWriter==3.2.9  # Excel write
Pillow>=10.1.0  # Image processing
python-magic>=0.4.27  # File type validation
xhtml2pdf>=0.2.11  # PDF generation

# Data Analysis
pandas==2.3.3
numpy==2.3.5

# HTTP & Communication
requests==2.32.5
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.3.0
frozenlist==1.8.0
multidict==6.7.0
yarl==1.20.1

# GeoIP
geoip2==5.2.0
maxminddb==2.8.2

# Security & Environment
python-dotenv>=1.0.0
cryptography>=41.0.7

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
asgiref==3.10.0
sqlparse==0.5.3
et-xmlfile==2.0.0
setuptools==80.9.0

# Production Server
gunicorn>=21.2.0

# Static Files
whitenoise>=6.6.0

# Monitoring (optional)
sentry-sdk>=1.39.1

# Admin Enhancements (optional)
django-import-export>=3.3.5

# ============================================================================
# Installation Instructions:
# 1. pip install -r requirements-complete.txt
# 2. python manage.py migrate
# 3. python manage.py createsuperuser
# 4. python manage.py runserver
# ============================================================================
